reservation/requirements.txt
grpcio
grpcio-tools
motor
beanie
loguru
PyYAML
kafka-python